requirements.txt
pydantic
tinydb
snoop
pyyaml